add session module and rename cli to client
