+++ client/pom.xml
@@ -28,7 +28,7 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
     <artifactId>iotdb-client</artifactId>
-    <name>IoTDB Client and Cli</name>
+    <name>IoTDB Client</name>
     <properties>
         <cli.test.skip>false</cli.test.skip>
         <cli.it.skip>${cli.test.skip}</cli.it.skip>
