+++ client/src/assembly/resources/sbin/start-client.sh
@@ -24,7 +24,7 @@ if [ -z "${IOTDB_CLI_HOME}" ]; then
 fi
 
 
-MAIN_CLASS=org.apache.iotdb.cli.Cli
+MAIN_CLASS=org.apache.iotdb.client.Client
 
 
 CLASSPATH=""
