+++ client/src/test/java/org/apache/iotdb/client/AbstractScript.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.cli;
+package org.apache.iotdb.client;
 
 import static org.junit.Assert.assertEquals;
