+++ docs/Documentation/UserGuide/6-JDBC API/1-JDBC API.md
@@ -19,7 +19,7 @@
 
 -->
 
-# Chaper6: JDBC API
+# Chapter6: JDBC API
 
 # Usage
 
@@ -57,7 +57,7 @@ This chapter provides an example of how to open a database connection, execute a
 
 Requires that you include the packages containing the JDBC classes needed for database programming.
 
-**NOTE: For faster insertion, the insertBatch() in Client is recommended.**
+**NOTE: For faster insertion, the insertBatch() in Session is recommended.**
 
 ```Java
 import java.sql.*;
