+++ example/pom.xml
@@ -36,7 +36,7 @@
     <modules>
         <module>kafka</module>
         <module>rocketmq</module>
-        <module>client</module>
+        <module>session</module>
         <module>tsfile</module>
     </modules>
     <build>
