+++ example/session/pom.xml
@@ -36,7 +36,7 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.iotdb</groupId>
-            <artifactId>iotdb-client</artifactId>
+            <artifactId>iotdb-session</artifactId>
             <version>0.9.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
