+++ server/src/main/antlr3/org/apache/iotdb/db/sql/parse/TSParser.g
@@ -119,7 +119,7 @@ ArrayList<ParseError> errors = new ArrayList<ParseError>();
 
     private static HashMap<String, String> xlateMap;
     static {
-        //this is used to support auto completion in CLI
+        //this is used to support auto completion in Client
         xlateMap = new HashMap<String, String>();
 
         // Keywords
